Refactor Satori into individual request files
Putting them all into a single file was getting a bit messy, so I wanted
to break it up a little.
